refactor: share gateway session store migration
(cherry picked from commit 198c2482eeb2133e78100b13753411d7c1fa0c29)
