+++ src/gateway/sessions-resolve.ts
@@ -10,7 +10,7 @@ import {
 import {
   listSessionsFromStore,
   loadCombinedSessionStoreForGateway,
-  pruneLegacyStoreKeys,
+  migrateAndPruneGatewaySessionStoreKey,
   resolveGatewaySessionStoreTarget,
 } from "./session-utils.js";
 
@@ -58,13 +58,10 @@ export async function resolveSessionKeyFromResolveParams(params: {
       };
     }
     await updateSessionStore(target.storePath, (s) => {
-      const liveTarget = resolveGatewaySessionStoreTarget({ cfg, key, store: s });
-      const canonicalKey = liveTarget.canonicalKey;
-      // Migrate the first legacy entry to the canonical key.
-      if (!s[canonicalKey] && s[legacyKey]) {
-        s[canonicalKey] = s[legacyKey];
+      const { primaryKey } = migrateAndPruneGatewaySessionStoreKey({ cfg, key, store: s });
+      if (!s[primaryKey] && s[legacyKey]) {
+        s[primaryKey] = s[legacyKey];
       }
-      pruneLegacyStoreKeys({ store: s, canonicalKey, candidates: liveTarget.storeKeys });
     });
     return { ok: true, key: target.canonicalKey };
   }
